Extract Coefficients in Original Coding

This extracts coefficients in terms of the original levels of the coefficients rather than the coded variable

# S3 method for class 'glmb'
dummy.coef(object, use.na = FALSE, ...)

# S3 method for class 'dummy_coef.glmb'
print(x, ...)

Arguments

object

a glmb model fit

use.na

logical flag for coefficients in a singular model. If use.na is true, undetermined coefficients will be missing; if false they will get one possible value.

x

object to be printed

...

arguments passed to or from other methods

Value

A list giving for each term the draws for the coefficients.

Details

A fitted linear model has coefficients for the contrasts of the factor terms, usually one less in number than the number of levels. This function re-expresses the coefficients in the original coding; as the coefficients will have been fitted in the reduced basis, any implied constraints (e.g., zero sum for contr.helmert or contr.sum) will be respected. There will be little point in using dummy.coef for contr.treatment contrasts, as the missing coefficients are by definition zero.

Examples

set.seed(333)
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
mysd<-1
mu<-matrix(0,5)
mu[1]=log(mean(counts))
V0<-((mysd)^2)*diag(5)
glmb.D93<-glmb(n=1000,counts ~ outcome + treatment, 
family = poisson(),pfamily=dNormal(mu=mu,Sigma=V0))
summary(glmb.D93)
#> Call
#> glmb(formula = counts ~ outcome + treatment, family = poisson(), 
#>     pfamily = dNormal(mu = mu, Sigma = V0), n = 1000)
#> 
#> Expected Deviance Residuals:
#>       1       2       3       4       5       6       7       8       9 
#> -0.5399  0.9667 -0.1405 -0.1038 -0.9671  1.0614  0.9687 -0.1031 -0.9567 
#> 
#> Prior and Maximum Likelihood Estimates with Standard Deviations
#> 
#>                          Prior.sd  Max Like. Like.sd
#> (Intercept)  2.813e+00  1.000e+00  3.045e+00   0.171
#> outcome2     0.000e+00  1.000e+00 -4.543e-01   0.202
#> outcome3     0.000e+00  1.000e+00 -2.930e-01   0.193
#> treatment2   0.000e+00  1.000e+00  1.217e-15   0.200
#> treatment3   0.000e+00  1.000e+00  8.438e-16   0.200
#> 
#> Bayesian Estimates Based on 1000 iid draws
#> 
#>             Post.Mode Post.Mean   Post.Sd MC Error Pr(tail)  
#> (Intercept)  3.027191  3.010036  0.170870        0   0.1279  
#> outcome2    -0.428954 -0.427228  0.203969        0   0.0160 *
#> outcome3    -0.272569 -0.271265  0.190823        0   0.0779 .
#> treatment2   0.004042  0.004123  0.198831        0   0.4885  
#> treatment3   0.004042  0.004311  0.190245        0   0.4925  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Distribution Percentiles
#> 
#>                  1.0%      2.5%      5.0%    Median     95.0%     97.5% 99.0%
#> (Intercept)  2.630959  2.678642  2.737872  3.007527  3.287809  3.332779 3.403
#> outcome2    -0.898758 -0.834953 -0.766190 -0.417820 -0.093401 -0.024640 0.027
#> outcome3    -0.716630 -0.643639 -0.586258 -0.264059  0.034240  0.102084 0.149
#> treatment2  -0.468938 -0.410120 -0.327144  0.004865  0.312898  0.386260 0.457
#> treatment3  -0.414533 -0.367326 -0.301737  0.002136  0.319114  0.371064 0.448
#> 
#> Effective Number of Parameters: 4.835259 
#> Expected Residual Deviance: 10.03303 
#> DIC: 56.50047 
#> 
#> Expected Mean dispersion: 1 
#> Sq.root of Expected Mean dispersion: 1 
#> 
#> Mean Likelihood Subgradient Candidates Per iid sample: 1.834 
#> 

myd=dummy.coef(glmb.D93)
print(myd)
#> 
#> Printing coefficients (first 5 records)
#> (Intercept) 
#> [1] 3.184155 3.007510 3.066190 2.966425 3.170668
#> outcome 
#>      1          2           3
#> [1,] 0 -0.4801134 -0.34835471
#> [2,] 0 -0.6649624 -0.31724672
#> [3,] 0 -0.1558581 -0.02023526
#> [4,] 0 -0.4057865 -0.25889315
#> [5,] 0 -0.4051943 -0.13224788
#> treatment 
#>      1           2           3
#> [1,] 0  0.06992985 -0.24241011
#> [2,] 0  0.25342902  0.21485606
#> [3,] 0 -0.26527971 -0.20025794
#> [4,] 0 -0.14138583  0.09427405
#> [5,] 0 -0.02621725 -0.31898199
#> coef.means: 
#> Full coefficients are 
#>                                                    
#> (Intercept):       3.010036                        
#> outcome:                  1           2           3
#>                   0.0000000  -0.4272279  -0.2712654
#> treatment:                1           2           3
#>                 0.000000000 0.004123343 0.004311096
#> coef.mode: 
#> Full coefficients are 
#>                                                    
#> (Intercept):       3.027191                        
#> outcome:                  1           2           3
#>                   0.0000000  -0.4289537  -0.2725692
#> treatment:                1           2           3
#>                 0.000000000 0.004041969 0.004041974