Bayesian Generalized Linear Models (iid Samples)

Generates iid samples for Bayesian Generalized Linear Models.

PackageOverview

Format

An object of class NULL of length 0.

Details

The glmbayes package produces iid samples for Bayesian Genereralized Linear Models and is intended as a Bayesian version of the glm function for classical models. Estimation can be performed using three main functions. For models with fixed dispersion parameters, the rglmb function is the workhorse function and comes with a minimialistic interface for the input and output. It is also suitable for use as part of block Gibbs sampling procedures. The glmb function is essentially a wrapper function for the rglmb function that provides an interface closer to that of the glm function. The rGamma_reg function can be leveraged in order to produce samples for the dispersion parameters associated with the gaussian and Gamma link functions. Most methods defined for the output of the glm function are also defined for the glmb, rglmb, and rGamma_reg functions (see their respective documentation for details).

For the regression parameters, multivariate normal priors are assumed. Simulation for the gaussian family with the identify link function is performed using standard procedures for multivariate normal densities. For all other families and link functions, simulation is performed using the likelihood subgradient approach of Nygren and Nygren (2006). This approach involves the construction of an enveloping function for the full posterior density followed by accept-reject sampling. For models that are approximately multivariate normal, the expected number of draws required per acceptance are bounded from above as noted in Nygren and Nygren (2006).

Currently implemented models include the gaussian (identity link), poisson/quasipoisson (log link), binomial/quasibinomial (logit, probit, and cloglog links), and Gamma (log link) families. These models all have log-concave likelihood functions that allow us to leverage the likelihood-subgradient approach for the iid sampling. Models that fail to have log-concave likelihood functions are not implemented. Our demos (viewable by entering the demo() command) provides examples of each of these families and links.

The current implementation requires separate use of the rGamma_reg function in order to generate samples for dispersion parameters (gaussian, Gammma, quasipoisson, quasi-binomial families). Our demos include examples of the joint use of the rglmb and rGamma_reg to produce samples for both regression and dispersion parameters using two-block Gibbs samplers. As these two-block Gibbs samplers likely are geometrically ergodic, future implementations may incorporate these two-block Gibbs samplers into the rglmb and glmb functions by leveraging theoretical bounds om convergence rates derived using Rosenthal (1996) type drift and minorization conditions.

The rglmb function can also be used in Block-Gibbs sampling implementations for Hierarchical Bayesian models. The demos associated with this package contains examples of such models.

References

Dobson, A. J. (1990) An Introduction to Generalized Linear Models. London: Chapman and Hall.

Hastie, T. J. and Pregibon, D. (1992) Generalized linear models. Chapter 6 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole. McCullagh P. and Nelder, J. A. (1989) Generalized Linear Models. London: Chapman and Hall.

Nygren, K.N. and Nygren, L.M (2006) Likelihood Subgradient Densities. Journal of the American Statistical Association. vol.101, no.475, pp 1144-1156.

Raiffa, Howard and Schlaifer, R (1961) Applied Statistical Decision Theory. Boston: Clinton Press, Inc.

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. New York: Springer.

Author

Kjell Nygren

Examples

## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
print(d.AD <- data.frame(treatment, outcome, counts))
#>   treatment outcome counts
#> 1         1       1     18
#> 2         1       2     17
#> 3         1       3     15
#> 4         2       1     20
#> 5         2       2     10
#> 6         2       3     20
#> 7         3       1     25
#> 8         3       2     13
#> 9         3       3     12

## Call to glm
glm.D93 <- glm(counts ~ outcome + treatment, 
               family = poisson())

## Using glmb
## Step 1: Set up Prior
ps=Prior_Setup(counts ~ outcome + treatment)
#> [1] "Variable names are:"
#> [1] "(Intercept)" "outcome2"    "outcome3"    "treatment2"  "treatment3" 
mu=ps$mu
V=ps$Sigma
# Step2A: Check the Prior
Prior_Check(counts ~ outcome + treatment,family = poisson(),
            pfamily=dNormal(mu=mu,Sigma=V))
#> [1] "Variables in the Model Are:"
#> [1] "(Intercept)" "outcome2"    "outcome3"    "treatment2"  "treatment3" 
#> [1] "At least one of the maximum likelihood estimates appears to be inconsistent with the prior"
#>                abs_ratio
#> (Intercept) 7.072923e+00
#> outcome2    2.297241e-01
#> outcome3    1.493395e-01
#> treatment2  6.168282e-16
#> treatment3  4.274795e-16
# Step2B: Update and Re-Check the Prior
mu[1,1]=log(mean(counts))
Prior_Check(counts ~ outcome + treatment,family = poisson(),
            pfamily=dNormal(mu=mu,Sigma=V))
#> [1] "Variables in the Model Are:"
#> [1] "(Intercept)" "outcome2"    "outcome3"    "treatment2"  "treatment3" 
#> [1] "The maximum likelihood estimates for all coefficients appear to be roughly consistent with the prior."
#>                abs_ratio
#> (Intercept) 1.199984e-01
#> outcome2    2.297241e-01
#> outcome3    1.493395e-01
#> treatment2  6.168282e-16
#> treatment3  4.274795e-16
# Step 3: Call the glmb function
glmb.D93<-glmb(counts ~ outcome + treatment, family=poisson(), 
               pfamily=dNormal(mu=mu,Sigma=V))

## ----Printed_Views------------------------------------------------------------
## Printed view of the output from the glm function 
print(glm.D93)
#> 
#> Call:  glm(formula = counts ~ outcome + treatment, family = poisson())
#> 
#> Coefficients:
#> (Intercept)     outcome2     outcome3   treatment2   treatment3  
#>   3.045e+00   -4.543e-01   -2.930e-01    1.218e-15    8.438e-16  
#> 
#> Degrees of Freedom: 8 Total (i.e. Null);  4 Residual
#> Null Deviance:	    10.58 
#> Residual Deviance: 5.129 	AIC: 56.76
## Printed view of the output from the glmb function 
print(glmb.D93)
#> 
#> Call:  glmb(formula = counts ~ outcome + treatment, family = poisson(), 
#>     pfamily = dNormal(mu = mu, Sigma = V))
#> 
#> Posterior Mean Coefficients:
#> (Intercept)     outcome2     outcome3   treatment2   treatment3  
#>    3.010238    -0.437015    -0.267092    -0.002132     0.009175  
#> 
#> Effective Number of Parameters: 4.709091 
#> Expected Residual Deviance: 9.915666 
#> DIC: 56.25693 
#> 

## ----Methods---------------------------------------------------------------
## Methods for class "lm"
methods(class="lm")
#>  [1] add1           addterm        alias          anova          boxcox        
#>  [6] case.names     coerce         confint        cooks.distance deviance      
#> [11] dfbeta         dfbetas        drop1          dropterm       dummy.coef    
#> [16] effects        extractAIC     family         formula        hatvalues     
#> [21] influence      initialize     kappa          labels         logLik        
#> [26] logtrans       model.frame    model.matrix   nobs           plot          
#> [31] predict        print          proj           qr             residuals     
#> [36] rstandard      rstudent       show           simulate       slotsFromS3   
#> [41] summary        variable.names vcov          
#> see '?methods' for accessing help and source code

## Methods for class "glm"
methods(class="glm")
#>  [1] add1           addterm        anova          coerce         confint       
#>  [6] cooks.distance deviance       drop1          dropterm       effects       
#> [11] extractAIC     family         formula        gamma.shape    influence     
#> [16] initialize     logLik         model.frame    nobs           predict       
#> [21] print          profile        residuals      rstandard      rstudent      
#> [26] show           sigma          slotsFromS3    summary        vcov          
#> [31] weights       
#> see '?methods' for accessing help and source code

## Methods for class "glmb"
methods(class="glmb")
#>  [1] anova          case.names     confint        cooks.distance dfbetas       
#>  [6] dummy.coef     extractAIC     influence      logLik         predict       
#> [11] print          residuals      rstandard      rstudent       simulate      
#> [16] summary        variable.names vcov          
#> see '?methods' for accessing help and source code

## ----summary--------------------------------------------------------------
## summary output for the "glm" class
summary(glm.D93)
#> 
#> Call:
#> glm(formula = counts ~ outcome + treatment, family = poisson())
#> 
#> Coefficients:
#>               Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  3.045e+00  1.709e-01  17.815   <2e-16 ***
#> outcome2    -4.543e-01  2.022e-01  -2.247   0.0246 *  
#> outcome3    -2.930e-01  1.927e-01  -1.520   0.1285    
#> treatment2   1.217e-15  2.000e-01   0.000   1.0000    
#> treatment3   8.438e-16  2.000e-01   0.000   1.0000    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> (Dispersion parameter for poisson family taken to be 1)
#> 
#>     Null deviance: 10.5814  on 8  degrees of freedom
#> Residual deviance:  5.1291  on 4  degrees of freedom
#> AIC: 56.761
#> 
#> Number of Fisher Scoring iterations: 4
#> 

## summary output for the "glm" class
summary(glmb.D93)
#> Call
#> glmb(formula = counts ~ outcome + treatment, family = poisson(), 
#>     pfamily = dNormal(mu = mu, Sigma = V))
#> 
#> Expected Deviance Residuals:
#>        1        2        3        4        5        6        7        8 
#> -0.53766  1.00515 -0.16012 -0.07745 -0.91092  1.07013  0.94558 -0.08463 
#>        9 
#> -0.99217 
#> 
#> Prior and Maximum Likelihood Estimates with Standard Deviations
#> 
#>                     mu   Prior.sd  Max Like. Like.sd
#> (Intercept)  2.813e+00  1.000e+00  3.045e+00   0.171
#> outcome2     0.000e+00  1.000e+00 -4.543e-01   0.202
#> outcome3     0.000e+00  1.000e+00 -2.930e-01   0.193
#> treatment2   0.000e+00  1.000e+00  1.217e-15   0.200
#> treatment3   0.000e+00  1.000e+00  8.438e-16   0.200
#> 
#> Bayesian Estimates Based on 1000 iid draws
#> 
#>             Post.Mode Post.Mean   Post.Sd MC Error Pr(tail)  
#> (Intercept)  3.027191  3.010238  0.158257        0   0.1049  
#> outcome2    -0.428954 -0.437015  0.190767        0   0.0120 *
#> outcome3    -0.272569 -0.267092  0.190689        0   0.0809 .
#> treatment2   0.004042 -0.002132  0.195116        0   0.4875  
#> treatment3   0.004042  0.009175  0.189146        0   0.4915  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Distribution Percentiles
#> 
#>                  1.0%      2.5%      5.0%    Median     95.0%     97.5% 99.0%
#> (Intercept)  2.639369  2.682230  2.744044  3.013577  3.272459  3.311820 3.340
#> outcome2    -0.884758 -0.807084 -0.752336 -0.432454 -0.124062 -0.067242 0.027
#> outcome3    -0.702670 -0.634801 -0.579518 -0.265546  0.046528  0.092623 0.209
#> treatment2  -0.436308 -0.384086 -0.331363 -0.003743  0.324383  0.400749 0.465
#> treatment3  -0.397736 -0.342898 -0.293665  0.006093  0.318721  0.377579 0.436
#> 
#> Effective Number of Parameters: 4.709091 
#> Expected Residual Deviance: 9.915666 
#> DIC: 56.25693 
#> 
#> Expected Mean dispersion: 1 
#> Sq.root of Expected Mean dispersion: 1 
#> 
#> Mean Likelihood Subgradient Candidates Per iid sample: 1.825 
#> 

## ----fitted outputs-------------------------------------------------------
## fitted outputs for the glm function
fitted(glm.D93)
#>        1        2        3        4        5        6        7        8 
#> 21.00000 13.33333 15.66667 21.00000 13.33333 15.66667 21.00000 13.33333 
#>        9 
#> 15.66667 

## ----glmb fitted outputs------------------------------------------------------
## mean of fitted outputs for the glm function
colMeans(fitted(glmb.D93))
#>        1        2        3        4        5        6        7        8 
#> 20.54580 13.35099 15.81351 20.54693 13.32072 15.77112 20.74691 13.46491 
#>        9 
#> 15.94853 

## ----predictions----------------------------------------------------------
## predictions for the glm function
predict(glm.D93)
#>        1        2        3        4        5        6        7        8 
#> 3.044522 2.590267 2.751535 3.044522 2.590267 2.751535 3.044522 2.590267 
#>        9 
#> 2.751535 

## predictions for the glmb function
colMeans(predict(glmb.D93)) 
#>        1        2        3        4        5        6        7        8 
#> 3.010238 2.573223 2.743146 3.008106 2.571091 2.741015 3.019413 2.582397 
#>        9 
#> 2.752321 

## ----residuals------------------------------------------------------------
## residuals for the glm function
residuals(glm.D93)
#>           1           2           3           4           5           6 
#> -0.67124923  0.96272360 -0.16964662 -0.21998507 -0.95552353  1.04938637 
#>           7           8           9 
#>  0.84715368 -0.09167147 -0.96656372 

## residuals for the glmb function
colMeans(residuals(glmb.D93))
#>           1           2           3           4           5           6 
#> -0.53766435  1.00515426 -0.16012027 -0.07744592 -0.91092255  1.07013232 
#>           7           8           9 
#>  0.94558007 -0.08463049 -0.99216789 

## ----vcov-----------------------------------------------------------------
## vcov for the glm function
vcov(glm.D93)
#>             (Intercept)      outcome2      outcome3    treatment2    treatment3
#> (Intercept)  0.02920635 -1.587302e-02 -1.587302e-02 -2.000000e-02 -2.000000e-02
#> outcome2    -0.01587302  4.087301e-02  1.587302e-02 -7.889764e-18 -7.435533e-18
#> outcome3    -0.01587302  1.587302e-02  3.714961e-02 -5.991636e-18 -6.847584e-18
#> treatment2  -0.02000000 -7.889764e-18 -5.991636e-18  4.000000e-02  2.000000e-02
#> treatment3  -0.02000000 -7.435533e-18 -6.847584e-18  2.000000e-02  4.000000e-02

## vcov for the glm function
vcov(glmb.D93)
#>             (Intercept)     outcome2     outcome3   treatment2   treatment3
#> (Intercept)  0.02504530 -0.012209317 -0.012869448 -0.016880247 -0.017377433
#> outcome2    -0.01220932  0.036391860  0.012284207 -0.002367430 -0.001246937
#> outcome3    -0.01286945  0.012284207  0.036362149 -0.002759632 -0.001186262
#> treatment2  -0.01688025 -0.002367430 -0.002759632  0.038070257  0.017413313
#> treatment3  -0.01737743 -0.001246937 -0.001186262  0.017413313  0.035776255

## ----confint--------------------------------------------------------------
## confint for the glm function
confint(glm.D93)
#> Waiting for profiling to be done...
#>                  2.5 %      97.5 %
#> (Intercept)  2.6958215  3.36655581
#> outcome2    -0.8577018 -0.06255840
#> outcome3    -0.6753696  0.08244089
#> treatment2  -0.3932548  0.39325483
#> treatment3  -0.3932548  0.39325483

## confint for the glm function
confint(glmb.D93)
#>                   2.5%       97.5%
#> (Intercept)  2.6822300  3.31182004
#> outcome2    -0.8070843 -0.06724181
#> outcome3    -0.6348014  0.09262313
#> treatment2  -0.3840862  0.40074879
#> treatment3  -0.3428980  0.37757878

## ----AIC/DIC------------------------------------------------------------------
## AIC for the glm function (equivalent degrees of freedom and the AIC)
extractAIC(glm.D93)
#> [1]  5.00000 56.76132

## DIC for the glmb function (estimated effective number of parameters and the DIC)
extractAIC(glmb.D93)
#>        pD       DIC 
#>  4.709091 56.256934 

## ----Deviance-------------------------------------------------------------
## Deviance for the glm function
deviance(glm.D93)
#> [1] 5.129141

## Deviance for the glmb function
mean(deviance(glmb.D93))
#> [1] 9.915666

## ----logLik---------------------------------------------------------------
## Deviance for the glm function
logLik(glm.D93)
#> 'log Lik.' -23.38066 (df=5)

## Deviance for the glmb function
mean(logLik(glmb.D93))
#> [1] -25.77392

## ----Model Frame----------------------------------------------------------
## Model Frame for the glm function
model.frame(glm.D93)
#>   counts outcome treatment
#> 1     18       1         1
#> 2     17       2         1
#> 3     15       3         1
#> 4     20       1         2
#> 5     10       2         2
#> 6     20       3         2
#> 7     25       1         3
#> 8     13       2         3
#> 9     12       3         3

## Model Frame for the glmb function
model.frame(glmb.D93$glm)
#>   counts outcome treatment
#> 1     18       1         1
#> 2     17       2         1
#> 3     15       3         1
#> 4     20       1         2
#> 5     10       2         2
#> 6     20       3         2
#> 7     25       1         3
#> 8     13       2         3
#> 9     12       3         3

## ----formula--------------------------------------------------------------
## formula for the glm function
formula(glm.D93)
#> counts ~ outcome + treatment
#> <environment: 0x000001b405f6f3f8>

## ----formula-------------------------------------------------------------
## formula for the glmb function
formula(glmb.D93)
#> counts ~ outcome + treatment
#> <environment: 0x000001b405f6f3f8>

## ----family--------------------------------------------------------------
## family for the glm function
family(glm.D93)
#> 
#> Family: poisson 
#> Link function: log 
#> 

## family for the glmb function
family(glmb.D93$glm)
#> 
#> Family: poisson 
#> Link function: log 
#> 

## ----nobs-----------------------------------------------------------------
## nobs for the glm function
nobs(glm.D93)
#> [1] 9

## nobs for the glmb function
nobs(glmb.D93)
#> [1] 9

## ----show-----------------------------------------------------------------
## show for the glm function
show(glm.D93)
#> 
#> Call:  glm(formula = counts ~ outcome + treatment, family = poisson())
#> 
#> Coefficients:
#> (Intercept)     outcome2     outcome3   treatment2   treatment3  
#>   3.045e+00   -4.543e-01   -2.930e-01    1.218e-15    8.438e-16  
#> 
#> Degrees of Freedom: 8 Total (i.e. Null);  4 Residual
#> Null Deviance:	    10.58 
#> Residual Deviance: 5.129 	AIC: 56.76

## show for the glmb function
show(glmb.D93)
#> 
#> Call:  glmb(formula = counts ~ outcome + treatment, family = poisson(), 
#>     pfamily = dNormal(mu = mu, Sigma = V))
#> 
#> Posterior Mean Coefficients:
#> (Intercept)     outcome2     outcome3   treatment2   treatment3  
#>    3.010238    -0.437015    -0.267092    -0.002132     0.009175  
#> 
#> Effective Number of Parameters: 4.709091 
#> Expected Residual Deviance: 9.915666 
#> DIC: 56.25693 
#>