The Bayesian Generalized Linear Model Dispersion Distribution

rGamma_reg is used to generate iid samples for the dispersion parameter in Bayesian Generalized Linear models. The model is specified by providing the model structure, regression coefficient, and a Gamma prior.

rGamma_reg(
  n,
  y,
  x,
  prior_list,
  offset = NULL,
  weights = 1,
  family = gaussian()
)

# S3 method for class 'rGamma_reg'
print(x, digits = max(3, getOption("digits") - 3), ...)

# S3 method for class 'rGamma_reg'
summary(object, ...)

Arguments

n

number of draws to generate. If length(n) > 1, the length is taken to be the number required.

y

a vector of observations of length m.

x

a design matrix of dimension m * p.

prior_list

a list with the prior parameters (shape and rate) for the prior distribution and a vector with the assumed value (beta) for the regression coefficients.

offset

this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases. One or more offset terms can be included in the formula instead or as well, and if more than one is specified their sum is used. See model.offset.

weights

an optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector.

family

a description of the error distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function. (See family for details of family functions.)

digits

Number of significant digits to use for printed outpute

object

an object of class "glmb_dispersion" that is to be summarized

...

further arguments passed to or from other methods

Value

rGamma_reg returns a object of class "rglmbdisp". The function summary (i.e., summary.rglmb) can be used to obtain or print a summary of the results. The generic accessor functions coefficients, fitted.values, residuals, and extractAIC can be used to extract various useful features of the value returned by rGamma_reg. An object of class "rglmbdisp" is a list containing at least the following components:

dispersion

an n by 1 matrix with simulated values for the dispersion

Prior

A list with two components. The first being the prior mean vector and the second the prior precision matrix

Details

The rGamma_reg function produces iid samples for the dispersion parameter in Bayesian generalized linear models (gaussian and Gamma families only). Core required inputs for the function include the data vector, the design matrix, an estimate for the regression coefficients, and a prior gamma distribution specification. The function returns the simulated Bayesian conditional dispersion and the prior specification. The most practical use of this function is likely to be in conjunction with the rglmb function as part of block Gibbs sampling procedues.

For the gaussian family, iid samples from the posterior density are genererated using standard simulation procedures for gamma distributions. For the Gamma family, , the samples are generated using accept-reject procedures by leveraging the likelihood subgradient approach of Nygren and Nygren (2006).

References

A reference

See also

Examples

## Annette Dobson (1990) "An Introduction to Generalized Linear Models".
## Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group,x=TRUE,y=TRUE)

lm_summary=summary(lm.D9)

lm_summary
#> 
#> Call:
#> lm(formula = weight ~ group, x = TRUE, y = TRUE)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -1.0710 -0.4938  0.0685  0.2462  1.3690 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   5.0320     0.2202  22.850 9.55e-15 ***
#> groupTrt     -0.3710     0.3114  -1.191    0.249    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 0.6964 on 18 degrees of freedom
#> Multiple R-squared:  0.07308,	Adjusted R-squared:  0.02158 
#> F-statistic: 1.419 on 1 and 18 DF,  p-value: 0.249
#> 
lm.D9$coefficients
#> (Intercept)    groupTrt 
#>       5.032      -0.371 

n<-10000
y<-lm.D9$y
x<-as.matrix(lm.D9$x)

### Set old regression and dispersion coefficients

b_old=lm.D9$coefficients
v_old=lm_summary$sigma^2

#### Set up for rglmb_dispersion

n0=0.1
shape=n0/2
rate=shape*v_old

rate/(shape-1)
#> [1] -0.02552412
rate/shape
#> [1] 0.4849583

#a1<-shape+n1/2
#b1<-rate+sum(SS)/2

#out<-1/rgamma(n,shape=a1,rate=b1)
## v0=sum(SS)/n1 ~ (2*rate+sum(SS))/(2*shape+n1)=[(2*rate+sum(SS))/2]/((2*shape+n1)/2) 
n1=length(y)
SS=v_old*n1

n1 # This is equal to 20
#> [1] 20

n0=2 # Prior observations
v_prior=v_old  # Prior point estimate for variance (the mean of (1/dispersion=1/v_prior))
wt0=(n0/n1)  

## set shape=0.01*(n1/2)
## set rate= 0.01*SS/2

shape=wt0*(n1/2)   ###  Shape is prior observations /2
rate=shape*v_prior  ### rate is essentiall prior SS - V in rmultireg should be this
rate/shape ## Should match v_prior (currently also v_old)
#> [1] 0.4849583

## We see that this currently matches 
### (test different v_prior with various prior observations below) 

prior_list=list(beta=b_old,shape=shape,rate=rate)

dispout<-rGamma_reg(n=n,y,x,prior_list=prior_list,
offset= rep(0, length(y)),family=gaussian())

mean(dispout$dispersion) 
#> [1] 0.4870285
v_prior
#> [1] 0.4849583
v_old
#> [1] 0.4849583
#summary(dispout)  ## Summary function not working - need to write

### Set up test rglmb regressions without and with prior for dispersion 

mu<-c(0,0)
mu=b_old  ### For testing purposes, set prior=b_old
P<-0.1*diag(2)
wt2=rep(1,length(y))

### Check
prior=list(mu=mu,Sigma=solve(P),dispersion=v_old)
outtemp1<-glmb(n = 1000, weight ~ group, family = gaussian(),
pfamily=dNormal(mu=mu,Sigma=solve(P),dispersion=v_old))
## Could use a residuals function here -- For now, maybe run the glmb function

summary(outtemp1)
#> Call
#> glmb(formula = weight ~ group, family = gaussian(), pfamily = dNormal(mu = mu, 
#>     Sigma = solve(P), dispersion = v_old), n = 1000)
#> 
#> Expected Deviance Residuals:
#>        1        2        3        4        5        6        7        8 
#> -0.86473  0.54527  0.14527  1.07527 -0.53473 -0.42473  0.13527 -0.50473 
#>        9       10       11       12       13       14       15       16 
#>  0.29527  0.10527  0.15734 -0.48266 -0.24266 -1.06266  1.21734 -0.82266 
#>       17       18       19       20 
#>  1.37734  0.23734 -0.33266  0.03734 
#> 
#> Prior and Maximum Likelihood Estimates with Standard Deviations
#> 
#>                    Prior.sd Max Like. Like.sd
#> (Intercept)  5.032    3.162     5.032   0.220
#> groupTrt    -0.371    3.162    -0.371   0.311
#> 
#> Bayesian Estimates Based on 1000 iid draws
#> 
#>             Post.Mode Post.Mean Post.Sd MC Error Pr(tail)
#> (Intercept)    5.0320    5.0347  0.2214        0    0.495
#> groupTrt      -0.3710   -0.3821  0.3030        0    0.473
#> 
#> Distribution Percentiles
#> 
#>                1.0%    2.5%    5.0%  Median   95.0%   97.5% 99.0%
#> (Intercept)  4.4953  4.5836  4.6486  5.0337  5.3932  5.4566 5.521
#> groupTrt    -1.1094 -0.9801 -0.8849 -0.3908  0.1123  0.2224 0.351
#> 
#> Effective Number of Parameters: 1.999052 
#> Expected Residual Deviance: 9.699477 
#> DIC: 21.4756 
#> 
#> Expected Mean dispersion: 0.4849583 
#> Sq.root of Expected Mean dispersion: 0.6963895 
#> 
#> Mean Likelihood Subgradient Candidates Per iid sample: 1 
#> 
mean(colMeans(residuals(outtemp1)^2))
#> [1] 0.4849739
v_old
#> [1] 0.4849583
colMeans((outtemp1$coefficients))
#> (Intercept)    groupTrt 
#>   5.0347321  -0.3820732 
b_old
#> (Intercept)    groupTrt 
#>       5.032      -0.371 
prior=list(mu=mu,P=P,dispersion=v_old)
outtemp2<-rglmb(n = 1000, y, x, family = gaussian(),
pfamily=dNormal(mu=mu,Sigma=solve(P),dispersion=v_old),
offset = rep(0, length(y)), weights = wt2)
summary(outtemp2)
#> Call
#> NULL
#> 
#> Prior Estimates with Standard Deviations
#> 
#>                    Prior.sd Approx.Prior.wt
#> (Intercept)  5.032    3.162           0.005
#> groupTrt    -0.371    3.162           0.010
#> 
#> Bayesian Estimates Based on 1000 iid draws
#> 
#>             Post.Mode Post.Mean Post.Sd MC Error Pr(tail)
#> (Intercept)    5.0320    5.0321  0.2183        0    0.492
#> groupTrt      -0.3710   -0.3685  0.3136        0    0.490
#> 
#> Distribution Percentiles
#> 
#>                1.0%    2.5%    5.0%  Median   95.0%   97.5% 99.0%
#> (Intercept)  4.5150  4.6117  4.6581  5.0369  5.4009  5.4783 5.563
#> groupTrt    -1.1237 -0.9950 -0.9113 -0.3607  0.1404  0.2309 0.338
colMeans((outtemp2$coefficients))
#> (Intercept)    groupTrt 
#>   5.0320973  -0.3685078 
b_old
#> (Intercept)    groupTrt 
#>       5.032      -0.371 

prior=list(mu=mu,Sigma=solve(P),shape=shape,rate=rate)
outtemp3<-glmb(n = 10000, weight ~ group,family = gaussian(),  
               dNormal_Gamma(mu=mu,Sigma=solve(P),shape=shape,rate=rate))
## Could use a residuals function here -- For now, maybe run the glmb function

summary(outtemp3)
#> Call
#> glmb(formula = weight ~ group, family = gaussian(), pfamily = dNormal_Gamma(mu = mu, 
#>     Sigma = solve(P), shape = shape, rate = rate), n = 10000)
#> 
#> Expected Deviance Residuals:
#>        1        2        3        4        5        6        7        8 
#> -0.86153  0.54847  0.14847  1.07847 -0.53153 -0.42153  0.13847 -0.50153 
#>        9       10       11       12       13       14       15       16 
#>  0.29847  0.10847  0.14843 -0.49157 -0.25157 -1.07157  1.20843 -0.83157 
#>       17       18       19       20 
#>  1.36843  0.22843 -0.34157  0.02843 
#> 
#> Prior and Maximum Likelihood Estimates with Standard Deviations
#> 
#>                    Prior.sd Max Like. Like.sd
#> (Intercept)  5.032    3.162     5.032   0.220
#> groupTrt    -0.371    3.162    -0.371   0.311
#> 
#> Bayesian Estimates Based on 10000 iid draws
#> 
#>             Post.Mode Post.Mean Post.Sd MC Error Pr(tail)
#> (Intercept)    5.0320    5.0315  0.2151        0    0.498
#> groupTrt      -0.3710   -0.3700  0.3066        0    0.497
#> 
#> Distribution Percentiles
#> 
#>                1.0%    2.5%    5.0%  Median   95.0%   97.5% 99.0%
#> (Intercept)  4.5183  4.6051  4.6801  5.0311  5.3827  5.4520 5.556
#> groupTrt    -1.0917 -0.9699 -0.8669 -0.3730  0.1334  0.2410 0.371
#> 
#> Effective Number of Parameters: 2.765901 
#> Expected Residual Deviance: 21.74862 
#> DIC: 45.8179 
#> 
#> Expected Mean dispersion: 0.4855357 
#> Sq.root of Expected Mean dispersion: 0.6968039 
#> 
#> Mean Likelihood Subgradient Candidates Per iid sample: 1 
#> 
mean(colMeans(residuals(outtemp3)^2))
#> [1] 0.4836226
v_old
#> [1] 0.4849583
colMeans((outtemp3$coefficients))
#> (Intercept)    groupTrt 
#>   5.0315328  -0.3699673 
b_old
#> (Intercept)    groupTrt 
#>       5.032      -0.371 
mean(outtemp3$dispersion)  ## Seems slightly smaller --> Needs qc
#> [1] 0.4855357
v_old
#> [1] 0.4849583