Setup Prior Objects

Sets up the structure for the Prior mean and Variance Matrices using information from a classical model.

Prior_Setup(
  formula,
  data = NULL,
  subset = NULL,
  na.action = na.fail,
  drop.unused.levels = FALSE,
  xlev = NULL,
  ...
)

Arguments

formula

a model formula or terms object or an R object.

data

a data frame, list or environment (or object coercible by as.data.frame to a data frame), containing the variables in formula. Neither a matrix nor an array will be accepted.

subset

a specification of the rows/observations to be used: defaults to all. This can be any valid indexing vector (see [.data.frame) for the rows of data, or a (logical) expression using variables in data or if that is not supplied, in formula. (See additional details about how this argument interacts with data-dependent bases under ‘Details’ below.)

na.action

how NAs are treated. The default is first, any na.action attribute of data, second a na.action setting of options, and third na.fail if that is unset. The factory-fresh default is na.omit. Another possible value is NULL.

drop.unused.levels

should factors have unused levels dropped? Defaults to FALSE.

xlev

a named list of character vectors giving the full set of levels to be assumed for each factor.

...

for model.frame methods, a mix of further arguments such as data, na.action, subset to pass to the default method. Any additional arguments (such as offset and weights or other named arguments) which reach the default method are used to create further columns in the model frame, with parenthesised names such as "(offset)".

For get_all_vars, further named columns to include in the model frame.

Value

A list with items related to the prior.

mu

An initial version of the prior mean vector, populated with all zeros

Sigma

An initial version of the prior Variance-Covariance vector, populated as the diagonal identity matrix

model

The model frame from object if it exists

x

The design matrix from object if it exists

See also

Examples

## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
print(d.AD <- data.frame(treatment, outcome, counts))
#>   treatment outcome counts
#> 1         1       1     18
#> 2         1       2     17
#> 3         1       3     15
#> 4         2       1     20
#> 5         2       2     10
#> 6         2       3     20
#> 7         3       1     25
#> 8         3       2     13
#> 9         3       3     12
## Step 1: Set up Prior
ps=Prior_Setup(counts ~ outcome + treatment)
#> [1] "Variable names are:"
#> [1] "(Intercept)" "outcome2"    "outcome3"    "treatment2"  "treatment3" 
mu=ps$mu
V=ps$Sigma
print(ps)
#> $mu
#>                   mu
#> (Intercept) 16.66667
#> outcome2     0.00000
#> outcome3     0.00000
#> treatment2   0.00000
#> treatment3   0.00000
#> 
#> $Sigma
#>             (Intercept) outcome2 outcome3 treatment2 treatment3
#> (Intercept)           1        0        0          0          0
#> outcome2              0        1        0          0          0
#> outcome3              0        0        1          0          0
#> treatment2            0        0        0          1          0
#> treatment3            0        0        0          0          1
#> 
#> $model
#>   counts outcome treatment
#> 1     18       1         1
#> 2     17       2         1
#> 3     15       3         1
#> 4     20       1         2
#> 5     10       2         2
#> 6     20       3         2
#> 7     25       1         3
#> 8     13       2         3
#> 9     12       3         3
#> 
#> $x
#>   (Intercept) outcome2 outcome3 treatment2 treatment3
#> 1           1        0        0          0          0
#> 2           1        1        0          0          0
#> 3           1        0        1          0          0
#> 4           1        0        0          1          0
#> 5           1        1        0          1          0
#> 6           1        0        1          1          0
#> 7           1        0        0          0          1
#> 8           1        1        0          0          1
#> 9           1        0        1          0          1
#> attr(,"assign")
#> [1] 0 1 1 2 2
#> attr(,"contrasts")
#> attr(,"contrasts")$outcome
#> [1] "contr.treatment"
#> 
#> attr(,"contrasts")$treatment
#> [1] "contr.treatment"
#> 
#>