Bayesian Regression Diagnostics

This function provides the basic quantities which are used in forming a wide variety of diagnostics for checking the quality of Bayesian regression fits.

glmb.influence.measures(model, infl = influence(model))

# S3 method for class 'glmb'
rstandard(model, ..., infl = influence(model))

# S3 method for class 'glmb'
rstudent(model, ..., infl = influence(model))

glmb.dffits(model, infl = influence(model))

# S3 method for class 'glmb'
dfbetas(model, ..., infl = influence(model))

glmb.covratio(model, infl = influence(model))

# S3 method for class 'glmb'
cooks.distance(model, ..., infl = influence(model))

Arguments

model

an R object, typically returned by lm or glm.

infl

influence structure as returned by influence.glmb (the latter only for the glm method of rstudent and cooks.distance).

...

further arguments passed to or from other methods.

Value

a list wih components:

Examples

set.seed(333)
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
mysd<-1
mu<-matrix(0,5)
mu[1]=log(mean(counts))
V0<-((mysd)^2)*diag(5)
glmb.D93<-glmb(counts ~ outcome + treatment, family = poisson(),pfamily=dNormal(mu=mu,Sigma=V0))

## Start setup here [First output from earlier optim optimization]
betastar=glmb.D93$coef.mode  # Posterior mode from optim
x=glmb.D93$x
y=glmb.D93$y
#offset=glmb.D93$offset   # not present in the output --> For now set to 0 vector
offset2=0*y   # Should return this from lower level functions
weights2=glmb.D93$prior.weights

## Check influence measures for original model
fit=glmb.wfit(x,y,weights2,offset2,family=poisson(),Bbar=mu,P=solve(V0),betastar)
influence.measures(fit)
#> Influence measures of
#> 	 NULL :
#> 
#>       dfb.1_  dfb.otc2  dfb.otc3  dfb.trt2  dfb.trt3     dffit cov.r   cook.d
#> 1  -1.318724  5.99e-01  6.28e-01  0.753402  0.753402 -1.318724 1.933 3.36e-01
#> 2   0.795705  1.27e+00 -9.70e-03 -1.001576 -1.001576  1.993449 0.449 5.91e-01
#> 3  -0.138294  1.14e-03 -1.99e-01  0.174801  0.174801 -0.331235 3.540 2.44e-02
#> 4  -0.117060  1.64e-01  1.72e-01 -0.213707  0.002200 -0.361356 4.154 2.91e-02
#> 5   0.358257 -1.24e+00  2.50e-02 -1.014497  0.025485 -1.985656 0.486 5.94e-01
#> 6  -0.508208 -2.98e-02  1.56e+00  1.426174 -0.035901  2.659421 0.214 8.93e-01
#> 7   0.883999 -1.24e+00 -1.30e+00 -0.016613  1.613845  2.728848 0.363 1.02e+00
#> 8   0.041628 -1.44e-01  2.91e-03  0.002961 -0.117879 -0.230723 3.487 1.19e-02
#> 9   0.423238  2.48e-02 -1.30e+00  0.029899 -1.187724 -2.214776 0.440 7.16e-01
#> 10 -0.044212  2.01e-02  2.11e-02  0.025259  0.025259 -0.044212 1.775 4.36e-04
#> 11 -0.049053  1.08e-01  4.28e-02 -0.000792 -0.000792  0.108065 1.570 2.54e-03
#> 12 -0.030840  2.57e-02  6.47e-02 -0.000498 -0.000498  0.064742 1.741 9.30e-04
#> 13  0.000564  7.23e-06  7.59e-06 -0.000987 -0.000475 -0.000987 1.873 2.19e-07
#> 14  0.000564  7.23e-06  7.59e-06 -0.000475 -0.000987 -0.000987 1.873 2.19e-07
#>       hat inf
#> 1  0.5673   *
#> 2  0.4923   *
#> 3  0.5218   *
#> 4  0.5900   *
#> 5  0.5000   *
#> 6  0.5306   *
#> 7  0.5900   *
#> 8  0.5000   *
#> 9  0.5306   *
#> 10 0.0275    
#> 11 0.0388    
#> 12 0.0355    
#> 13 0.0378    
#> 14 0.0378    

print(fit)
#> 
#> Call:
#> NULL
#> 
#> Coefficients:
#> (Intercept)     outcome2     outcome3   treatment2   treatment3  
#>    3.027191    -0.428953    -0.272569     0.004042     0.004042  
#> 
print(glmb.D93$coef.mode)
#> [1]  3.027191119 -0.428953693 -0.272569221  0.004041969  0.004041974

### Now try a strong prior with poorly chosen intercept
mu1=0*mu
V1=0.1*V0
glmb2.D93<-glmb(counts ~ outcome + treatment, family = poisson(),pfamily=dNormal(mu=mu1,Sigma=V1))

Bbar2=mu1  # Prior mean
betastar2=glmb2.D93$coef.mode  # Posterior mode from optim
fit2=glmb.wfit(x,y,weights2,offset2,family=poisson(),Bbar2,P=solve(V1),betastar2)

influence.measures(fit2)
#> Influence measures of
#> 	 NULL :
#> 
#>      dfb.1_ dfb.otc2 dfb.otc3  dfb.trt2  dfb.trt3    dffit    cov.r   cook.d
#> 1   0.48766 -0.20271 -0.21342 -0.240330 -0.240332  0.48766 1.54e+00 4.90e-02
#> 2   0.25187  0.34436 -0.02125 -0.268551 -0.268553  0.55427 1.83e+00 6.40e-02
#> 3   0.11359 -0.00723  0.14755 -0.124322 -0.124323  0.24705 2.57e+00 1.35e-02
#> 4   0.18233 -0.19873 -0.20922  0.259999 -0.019529  0.44148 2.50e+00 4.23e-02
#> 5   0.01203 -0.28107  0.05047 -0.243283  0.050505 -0.50068 2.78e+00 5.44e-02
#> 6  -0.01170 -0.02982  0.17499  0.164955 -0.034869  0.32841 3.28e+00 2.39e-02
#> 7   0.38365 -0.41815 -0.44024 -0.041091  0.547077  0.92895 1.41e+00 1.67e-01
#> 8   0.00442 -0.10329  0.01855  0.018560 -0.089404 -0.18399 3.25e+00 7.58e-03
#> 9   0.01849  0.04714 -0.27655  0.055108 -0.260697 -0.51903 2.95e+00 5.85e-02
#> 10 -4.69067  1.94983  2.05283  2.311656  2.311671 -4.69067 2.11e-05 4.85e-01
#> 11 -0.00315  0.00758  0.00262 -0.000464 -0.000464  0.00758 2.56e+00 1.29e-05
#> 12  0.03962 -0.03131 -0.09053  0.005832  0.005832 -0.09053 2.48e+00 1.84e-03
#> 13  0.12891  0.01601  0.01685 -0.261580 -0.103342 -0.26158 2.30e+00 1.51e-02
#> 14  0.12891  0.01601  0.01685 -0.103342 -0.261578 -0.26158 2.30e+00 1.51e-02
#>      hat inf
#> 1  0.244    
#> 2  0.326    
#> 3  0.347    
#> 4  0.398    
#> 5  0.459   *
#> 6  0.487   *
#> 7  0.398    
#> 8  0.459   *
#> 9  0.487   *
#> 10 0.230   *
#> 11 0.297    
#> 12 0.282    
#> 13 0.292    
#> 14 0.292    

print(fit2)
#> 
#> Call:
#> NULL
#> 
#> Coefficients:
#> (Intercept)     outcome2     outcome3   treatment2   treatment3  
#>    2.363486    -0.009802     0.122589     0.340056     0.340056  
#> 
print(glmb2.D93$coef.mode)
#> [1]  2.363485555 -0.009797081  0.122587885  0.340049094  0.340066361

influence(glmb2.D93)
#> $hat
#>  [1] 0.2444410 0.3261137 0.3472719 0.3981078 0.4590068 0.4873046 0.3981108
#>  [8] 0.4590107 0.4873085 0.2299986 0.2972437 0.2819494 0.2920672 0.2920654
#> 
#> $coefficients
#>        (Intercept)     outcome2     outcome3    treatment2   treatment3
#>  [1,]  0.224411797 -0.106047652 -0.108739010 -0.1246270805 -0.124627461
#>  [2,]  0.116574001  0.181193322 -0.010889800 -0.1400676089 -0.140068036
#>  [3,]  0.054232896 -0.003922223  0.077998224 -0.0668881167 -0.066888321
#>  [4,]  0.086102519 -0.106686056 -0.109393616  0.1383594288 -0.010392459
#>  [5,]  0.005679600 -0.150887167  0.026385000 -0.1294578825  0.026875121
#>  [6,] -0.005584718 -0.016189238  0.092508268  0.0887561061 -0.018761790
#>  [7,]  0.171058886 -0.211952863 -0.217331965 -0.0206460063  0.274874462
#>  [8,]  0.002120262 -0.056324513  0.009849251  0.0100320421 -0.048324469
#>  [9,]  0.008735668  0.025322331 -0.144696084  0.0293456930 -0.138825075
#> [10,] -0.705971062  0.333612468  0.342079140  0.3920609940  0.392062190
#> [11,] -0.001516008  0.004146042  0.001396417 -0.0002514792 -0.000251480
#> [12,]  0.019026658 -0.017091950 -0.048135857  0.0031561892  0.003156199
#> [13,]  0.061355020  0.008660426  0.008880217 -0.1402945633 -0.055425896
#> [14,]  0.061355050  0.008660430  0.008880221 -0.0554257549 -0.140293337
#> 
#> $sigma
#>  [1] 3.0343276 3.0518910 3.1481778 3.1138402 3.1136942 3.1484196 2.9399904
#>  [8] 3.1627920 3.1159546 0.9924051 3.1706425 3.1665412 3.1383056 3.1383057
#> 
#> $wt.res
#>  [1]  2.26133260  1.99614294  0.86147034  1.31137191 -1.24484255  0.75939034
#>  [7]  2.60526193 -0.46467873 -1.18777571 -7.47400044  0.03099743 -0.38766144
#> [13] -1.07535074 -1.07535074
#> 
influence(glmb2.D93,do.coef=TRUE)
#> $hat
#>  [1] 0.2444410 0.3261137 0.3472719 0.3981078 0.4590068 0.4873046 0.3981108
#>  [8] 0.4590107 0.4873085 0.2299986 0.2972437 0.2819494 0.2920672 0.2920654
#> 
#> $coefficients
#>        (Intercept)     outcome2     outcome3    treatment2   treatment3
#>  [1,]  0.224411797 -0.106047652 -0.108739010 -0.1246270805 -0.124627461
#>  [2,]  0.116574001  0.181193322 -0.010889800 -0.1400676089 -0.140068036
#>  [3,]  0.054232896 -0.003922223  0.077998224 -0.0668881167 -0.066888321
#>  [4,]  0.086102519 -0.106686056 -0.109393616  0.1383594288 -0.010392459
#>  [5,]  0.005679600 -0.150887167  0.026385000 -0.1294578825  0.026875121
#>  [6,] -0.005584718 -0.016189238  0.092508268  0.0887561061 -0.018761790
#>  [7,]  0.171058886 -0.211952863 -0.217331965 -0.0206460063  0.274874462
#>  [8,]  0.002120262 -0.056324513  0.009849251  0.0100320421 -0.048324469
#>  [9,]  0.008735668  0.025322331 -0.144696084  0.0293456930 -0.138825075
#> [10,] -0.705971062  0.333612468  0.342079140  0.3920609940  0.392062190
#> [11,] -0.001516008  0.004146042  0.001396417 -0.0002514792 -0.000251480
#> [12,]  0.019026658 -0.017091950 -0.048135857  0.0031561892  0.003156199
#> [13,]  0.061355020  0.008660426  0.008880217 -0.1402945633 -0.055425896
#> [14,]  0.061355050  0.008660430  0.008880221 -0.0554257549 -0.140293337
#> 
#> $sigma
#>  [1] 3.0343276 3.0518910 3.1481778 3.1138402 3.1136942 3.1484196 2.9399904
#>  [8] 3.1627920 3.1159546 0.9924051 3.1706425 3.1665412 3.1383056 3.1383057
#> 
#> $wt.res
#>  [1]  2.26133260  1.99614294  0.86147034  1.31137191 -1.24484255  0.75939034
#>  [7]  2.60526193 -0.46467873 -1.18777571 -7.47400044  0.03099743 -0.38766144
#> [13] -1.07535074 -1.07535074
#> 
influence(glmb2.D93,do.coef=FALSE)
#> $hat
#>  [1] 0.2444410 0.3261137 0.3472719 0.3981078 0.4590068 0.4873046 0.3981108
#>  [8] 0.4590107 0.4873085 0.2299986 0.2972437 0.2819494 0.2920672 0.2920654
#> 
#> $<NA>
#> NULL
#> 
#> $sigma
#>  [1] 3.0343276 3.0518910 3.1481778 3.1138402 3.1136942 3.1484196 2.9399904
#>  [8] 3.1627920 3.1159546 0.9924051 3.1706425 3.1665412 3.1383056 3.1383057
#> 
#> $wt.res
#>  [1]  2.26133260  1.99614294  0.86147034  1.31137191 -1.24484255  0.75939034
#>  [7]  2.60526193 -0.46467873 -1.18777571 -7.47400044  0.03099743 -0.38766144
#> [13] -1.07535074 -1.07535074
#> 

#glmb2.D93$qr=glmb2.D93$fit$qr
#influence.measures(glmb2.D93,influence(glmb2.D93))
#influence.measures(glmb2.D93$fit,influence(glmb2.D93))
#influence.measures(glmb2.D93$fit,influence(glmb2.D93,do.coef=FALSE))
glmb.influence.measures(glmb2.D93,influence(glmb2.D93))
#> Influence measures of
#> 	 NULL :
#> 
#>      dfb.1_ dfb.otc2 dfb.otc3  dfb.trt2  dfb.trt3    dffit    cov.r   cook.d
#> 1   0.48766 -0.20271 -0.21342 -0.240330 -0.240332  0.48766 1.54e+00 4.90e-02
#> 2   0.25187  0.34436 -0.02125 -0.268551 -0.268553  0.55427 1.83e+00 6.40e-02
#> 3   0.11359 -0.00723  0.14755 -0.124322 -0.124323  0.24705 2.57e+00 1.35e-02
#> 4   0.18233 -0.19873 -0.20922  0.259999 -0.019529  0.44148 2.50e+00 4.23e-02
#> 5   0.01203 -0.28107  0.05047 -0.243283  0.050505 -0.50068 2.78e+00 5.44e-02
#> 6  -0.01170 -0.02982  0.17499  0.164955 -0.034869  0.32841 3.28e+00 2.39e-02
#> 7   0.38365 -0.41815 -0.44024 -0.041091  0.547077  0.92895 1.41e+00 1.67e-01
#> 8   0.00442 -0.10329  0.01855  0.018560 -0.089404 -0.18399 3.25e+00 7.58e-03
#> 9   0.01849  0.04714 -0.27655  0.055108 -0.260697 -0.51903 2.95e+00 5.85e-02
#> 10 -4.69067  1.94983  2.05283  2.311656  2.311671 -4.69067 2.11e-05 4.85e-01
#> 11 -0.00315  0.00758  0.00262 -0.000464 -0.000464  0.00758 2.56e+00 1.29e-05
#> 12  0.03962 -0.03131 -0.09053  0.005832  0.005832 -0.09053 2.48e+00 1.84e-03
#> 13  0.12891  0.01601  0.01685 -0.261580 -0.103342 -0.26158 2.30e+00 1.51e-02
#> 14  0.12891  0.01601  0.01685 -0.103342 -0.261578 -0.26158 2.30e+00 1.51e-02
#>      hat inf
#> 1  0.244    
#> 2  0.326    
#> 3  0.347    
#> 4  0.398    
#> 5  0.459   *
#> 6  0.487   *
#> 7  0.398    
#> 8  0.459   *
#> 9  0.487   *
#> 10 0.230   *
#> 11 0.297    
#> 12 0.282    
#> 13 0.292    
#> 14 0.292    

## Do not need method functions
dfbeta(glmb2.D93,influence(glmb2.D93))
#>        (Intercept)     outcome2     outcome3    treatment2   treatment3
#>  [1,]  0.224411797 -0.106047652 -0.108739010 -0.1246270805 -0.124627461
#>  [2,]  0.116574001  0.181193322 -0.010889800 -0.1400676089 -0.140068036
#>  [3,]  0.054232896 -0.003922223  0.077998224 -0.0668881167 -0.066888321
#>  [4,]  0.086102519 -0.106686056 -0.109393616  0.1383594288 -0.010392459
#>  [5,]  0.005679600 -0.150887167  0.026385000 -0.1294578825  0.026875121
#>  [6,] -0.005584718 -0.016189238  0.092508268  0.0887561061 -0.018761790
#>  [7,]  0.171058886 -0.211952863 -0.217331965 -0.0206460063  0.274874462
#>  [8,]  0.002120262 -0.056324513  0.009849251  0.0100320421 -0.048324469
#>  [9,]  0.008735668  0.025322331 -0.144696084  0.0293456930 -0.138825075
#> [10,] -0.705971062  0.333612468  0.342079140  0.3920609940  0.392062190
#> [11,] -0.001516008  0.004146042  0.001396417 -0.0002514792 -0.000251480
#> [12,]  0.019026658 -0.017091950 -0.048135857  0.0031561892  0.003156199
#> [13,]  0.061355020  0.008660426  0.008880217 -0.1402945633 -0.055425896
#> [14,]  0.061355050  0.008660430  0.008880221 -0.0554257549 -0.140293337
dfbeta(glmb2.D93$fit,influence(glmb2.D93))
#>        (Intercept)     outcome2     outcome3    treatment2   treatment3
#>  [1,]  0.224411797 -0.106047652 -0.108739010 -0.1246270805 -0.124627461
#>  [2,]  0.116574001  0.181193322 -0.010889800 -0.1400676089 -0.140068036
#>  [3,]  0.054232896 -0.003922223  0.077998224 -0.0668881167 -0.066888321
#>  [4,]  0.086102519 -0.106686056 -0.109393616  0.1383594288 -0.010392459
#>  [5,]  0.005679600 -0.150887167  0.026385000 -0.1294578825  0.026875121
#>  [6,] -0.005584718 -0.016189238  0.092508268  0.0887561061 -0.018761790
#>  [7,]  0.171058886 -0.211952863 -0.217331965 -0.0206460063  0.274874462
#>  [8,]  0.002120262 -0.056324513  0.009849251  0.0100320421 -0.048324469
#>  [9,]  0.008735668  0.025322331 -0.144696084  0.0293456930 -0.138825075
#> [10,] -0.705971062  0.333612468  0.342079140  0.3920609940  0.392062190
#> [11,] -0.001516008  0.004146042  0.001396417 -0.0002514792 -0.000251480
#> [12,]  0.019026658 -0.017091950 -0.048135857  0.0031561892  0.003156199
#> [13,]  0.061355020  0.008660426  0.008880217 -0.1402945633 -0.055425896
#> [14,]  0.061355050  0.008660430  0.008880221 -0.0554257549 -0.140293337
hatvalues(glmb2.D93,influence(glmb2.D93))
#>  [1] 0.2444410 0.3261137 0.3472719 0.3981078 0.4590068 0.4873046 0.3981108
#>  [8] 0.4590107 0.4873085 0.2299986 0.2972437 0.2819494 0.2920672 0.2920654
hatvalues(glmb2.D93$fit,influence(glmb2.D93))
#>  [1] 0.2444410 0.3261137 0.3472719 0.3981078 0.4590068 0.4873046 0.3981108
#>  [8] 0.4590107 0.4873085 0.2299986 0.2972437 0.2819494 0.2920672 0.2920654

# Methods (implmented)

rstandard(glmb2.D93,infl=influence(glmb2.D93))
#>  [1]  0.8702731  0.8134364  0.3566970  0.5654469 -0.5661660  0.3547813
#>  [7]  1.1233586 -0.2113410 -0.5549216 -2.8492639  0.0123694 -0.1530384
#> [13] -0.4275426 -0.4275421
rstandard(glmb2.D93)
#>  [1]  0.8702731  0.8134364  0.3566970  0.5654469 -0.5661660  0.3547813
#>  [7]  1.1233586 -0.2113410 -0.5549216 -2.8492639  0.0123694 -0.1530384
#> [13] -0.4275426 -0.4275421

# Need Metods

## Needs a method function
#dfbetas(glmb2.D93,influence(glmb2.D93))
dfbetas(glmb2.D93)
#>        (Intercept)     outcome2     outcome3    treatment2    treatment3
#>  [1,]  0.487663863 -0.202713301 -0.213421085 -0.2403303236 -0.2403318006
#>  [2,]  0.251866292  0.344363277 -0.021250312 -0.2685513379 -0.2685529883
#>  [3,]  0.113590211 -0.007226311  0.147550227 -0.1243221005 -0.1243228645
#>  [4,]  0.182329482 -0.198726141 -0.209223314  0.2599986322 -0.0195290885
#>  [5,]  0.012027606 -0.281073566  0.050465613 -0.2432826631  0.0505050064
#>  [6,] -0.011696234 -0.029824794  0.174985621  0.1649545337 -0.0348691713
#>  [7,]  0.383651673 -0.418154757 -0.440242655 -0.0410911950  0.5470769954
#>  [8,]  0.004420345 -0.103292902  0.018545861  0.0185599731 -0.0894038938
#>  [9,]  0.018485975  0.047136380 -0.276554087  0.0551076484 -0.2606974420
#> [10,] -4.690674224  1.949830875  2.052825436  2.3116563291  2.3116705356
#> [11,] -0.003152766  0.007584554  0.002622903 -0.0004641021 -0.0004641049
#> [12,]  0.039620024 -0.031307625 -0.090531139  0.0058322551  0.0058322910
#> [13,]  0.128911668  0.016006179  0.016851663 -0.2615798167 -0.1033421421
#> [14,]  0.128911729  0.016006187  0.016851671 -0.1033415557 -0.2615783324
dfbetas(glmb2.D93,influence(glmb2.D93,do.coef=TRUE))
#>        (Intercept)     outcome2     outcome3    treatment2    treatment3
#>  [1,]  0.487663863 -0.202713301 -0.213421085 -0.2403303236 -0.2403318006
#>  [2,]  0.251866292  0.344363277 -0.021250312 -0.2685513379 -0.2685529883
#>  [3,]  0.113590211 -0.007226311  0.147550227 -0.1243221005 -0.1243228645
#>  [4,]  0.182329482 -0.198726141 -0.209223314  0.2599986322 -0.0195290885
#>  [5,]  0.012027606 -0.281073566  0.050465613 -0.2432826631  0.0505050064
#>  [6,] -0.011696234 -0.029824794  0.174985621  0.1649545337 -0.0348691713
#>  [7,]  0.383651673 -0.418154757 -0.440242655 -0.0410911950  0.5470769954
#>  [8,]  0.004420345 -0.103292902  0.018545861  0.0185599731 -0.0894038938
#>  [9,]  0.018485975  0.047136380 -0.276554087  0.0551076484 -0.2606974420
#> [10,] -4.690674224  1.949830875  2.052825436  2.3116563291  2.3116705356
#> [11,] -0.003152766  0.007584554  0.002622903 -0.0004641021 -0.0004641049
#> [12,]  0.039620024 -0.031307625 -0.090531139  0.0058322551  0.0058322910
#> [13,]  0.128911668  0.016006179  0.016851663 -0.2615798167 -0.1033421421
#> [14,]  0.128911729  0.016006187  0.016851671 -0.1033415557 -0.2615783324
dfbetas(glmb2.D93$fit,influence(glmb2.D93))
#>        (Intercept)     outcome2     outcome3    treatment2    treatment3
#>  [1,]  0.487663863 -0.202713301 -0.213421085 -0.2403303236 -0.2403318006
#>  [2,]  0.251866292  0.344363277 -0.021250312 -0.2685513379 -0.2685529883
#>  [3,]  0.113590211 -0.007226311  0.147550227 -0.1243221005 -0.1243228645
#>  [4,]  0.182329482 -0.198726141 -0.209223314  0.2599986322 -0.0195290885
#>  [5,]  0.012027606 -0.281073566  0.050465613 -0.2432826631  0.0505050064
#>  [6,] -0.011696234 -0.029824794  0.174985621  0.1649545337 -0.0348691713
#>  [7,]  0.383651673 -0.418154757 -0.440242655 -0.0410911950  0.5470769954
#>  [8,]  0.004420345 -0.103292902  0.018545861  0.0185599731 -0.0894038938
#>  [9,]  0.018485975  0.047136380 -0.276554087  0.0551076484 -0.2606974420
#> [10,] -4.690674224  1.949830875  2.052825436  2.3116563291  2.3116705356
#> [11,] -0.003152766  0.007584554  0.002622903 -0.0004641021 -0.0004641049
#> [12,]  0.039620024 -0.031307625 -0.090531139  0.0058322551  0.0058322910
#> [13,]  0.128911668  0.016006179  0.016851663 -0.2615798167 -0.1033421421
#> [14,]  0.128911729  0.016006187  0.016851671 -0.1033415557 -0.2615783324

# Needs a method function
cooks.distance(glmb2.D93)
#>  [1] 4.900571e-02 6.404122e-02 1.353837e-02 4.229564e-02 5.439324e-02
#>  [6] 2.392723e-02 1.669376e-01 7.579345e-03 5.853844e-02 4.849859e-01
#> [11] 1.294301e-05 1.839276e-03 1.508271e-02 1.508254e-02
cooks.distance(glmb2.D93$fit,influence(glmb2.D93))
#>  [1] 4.900571e-02 6.404122e-02 1.353837e-02 4.229564e-02 5.439324e-02
#>  [6] 2.392723e-02 1.669376e-01 7.579345e-03 5.853844e-02 4.849859e-01
#> [11] 1.294301e-05 1.839276e-03 1.508271e-02 1.508254e-02

# Needs a method function (now works)
#rstandard(glmb2.D93$fit,influence(glmb2.D93))

# Needs a method function
rstudent(glmb2.D93)
#>  [1]  0.85736899  0.79676325  0.33869979  0.54283794 -0.54355382  0.33685481
#>  [7]  1.14221329 -0.19975048 -0.53237206 -8.58259030  0.01166208 -0.14447408
#> [13] -0.40724792 -0.40724739
rstudent(glmb2.D93$fit,influence(glmb2.D93))
#>  [1]  0.85736899  0.79676325  0.33869979  0.54283794 -0.54355382  0.33685481
#>  [7]  1.14221329 -0.19975048 -0.53237206 -8.58259030  0.01166208 -0.14447408
#> [13] -0.40724792 -0.40724739

# Not a method, separate function
glmb.dffits(glmb2.D93)
#>  [1]  0.487663863  0.554268853  0.247049279  0.441480129 -0.500675335
#>  [6]  0.328407686  0.928946937 -0.183994536 -0.519026058 -4.690674224
#> [11]  0.007584554 -0.090531139 -0.261579817 -0.261578332
dffits(glmb2.D93$fit,influence(glmb2.D93)) 
#>  [1]  0.487663863  0.554268853  0.247049279  0.441480129 -0.500675335
#>  [6]  0.328407686  0.928946937 -0.183994536 -0.519026058 -4.690674224
#> [11]  0.007584554 -0.090531139 -0.261579817 -0.261578332

# Not a method - separate function
glmb.covratio(glmb2.D93)  
#>  [1] 1.536772e+00 1.825394e+00 2.571066e+00 2.498604e+00 2.778566e+00
#>  [6] 3.275817e+00 1.406681e+00 3.249156e+00 2.953303e+00 2.111793e-05
#> [11] 2.564017e+00 2.477133e+00 2.297277e+00 2.297271e+00
covratio(glmb2.D93$fit,influence(glmb2.D93))  
#>  [1] 1.536772e+00 1.825394e+00 2.571066e+00 2.498604e+00 2.778566e+00
#>  [6] 3.275817e+00 1.406681e+00 3.249156e+00 2.953303e+00 2.111793e-05
#> [11] 2.564017e+00 2.477133e+00 2.297277e+00 2.297271e+00

# Needs a method function but requires a different approach
# The influence function actually stores this measure
## This seems to require more work to create a method function
## Should 
#hat(glmb2.D93,influence(glmb2.D93))
#hat(glmb2.D93$fit,influence(glmb2.D93))

infl=influence(glmb2.D93)
hat2=infl$hat               
hat2               
#>  [1] 0.2444410 0.3261137 0.3472719 0.3981078 0.4590068 0.4873046 0.3981108
#>  [8] 0.4590107 0.4873085 0.2299986 0.2972437 0.2819494 0.2920672 0.2920654