Return family functions used during simulation and post processing

This function takes as input a family object and returns a set of functions that are used during simulation and summarization of models using the glmb, and rglmb functions.

glmbfamfunc(family)

# S3 method for class 'glmbfamfunc'
print(x, ...)

Arguments

family

an object of class family

x

an object of class "glmbfamfunc" for which a printed output is desired.

...

additional optional arguments

Value

A list with the following components

f1

Negative log-likelihood function

f2

Negative log-posterior function

f3

Gradient function for negative log-posterior function

f4

Deviance function

f7

Another function

Details

This function takes as input a family and returns a set of functions related to the family.

Examples

famfunc<-glmbfamfunc(binomial(logit))

print(famfunc)
#> $f1
#> function (b, y, x, alpha = 0, wt = 1) 
#> {
#>     lambda <- t(exp(alpha + x %*% b))
#>     p <- lambda/(1 + lambda)
#>     -sum(dbinom(round(wt * y), round(wt), p, log = TRUE))
#> }
#> <bytecode: 0x000001b40443d700>
#> <environment: 0x000001b407281af8>
#> 
#> $f2
#> function (b, y, x, mu, P, alpha = 0, wt = 1) 
#> {
#>     lambda <- t(exp(alpha + x %*% b))
#>     p <- lambda/(1 + lambda)
#>     -sum(dbinom(round(wt * y), round(wt), p, log = TRUE)) + 0.5 * 
#>         t((b - mu)) %*% P %*% (b - mu)
#> }
#> <bytecode: 0x000001b4044047b0>
#> <environment: 0x000001b407281af8>
#> 
#> $f3
#> function (b, y, x, mu, P, alpha = 0, wt = 1) 
#> {
#>     p <- 1/(1 + t(exp(-alpha - x %*% b)))
#>     t(x) %*% ((t(p) - y) * wt) + P %*% (b - mu)
#> }
#> <bytecode: 0x000001b404405ab8>
#> <environment: 0x000001b407281af8>
#> 
#> $f4
#> function (b, y, x, alpha = 0, wt = 1, dispersion = 1) 
#> {
#>     (2 * f1(b, y, x, alpha, wt/dispersion) + 2 * sum(dbinom(round((wt/dispersion) * 
#>         y), round(wt/dispersion), y, log = TRUE)))
#> }
#> <bytecode: 0x000001b404428a88>
#> <environment: 0x000001b407281af8>
#> 
#> $f7
#> function (b, y, x, mu, P, alpha = 0, wt = 1) 
#> {
#>     l2 <- length(y)
#>     l1 <- length(b)
#>     ltemp <- length(wt)
#>     if (ltemp == 1) {
#>         Ptemp <- wt * diag(l2)
#>     }
#>     else {
#>         Ptemp <- diag(wt)
#>     }
#>     p <- 1/(1 + t(exp(-alpha - x %*% b)))
#>     Pout <- matrix(0, nrow = l1, ncol = l1)
#>     for (i in 1:l2) {
#>         Pout <- Pout + Ptemp[i, i] * p[i] * (1 - p[i]) * x[i, 
#>             ] %*% t(x[i, ])
#>     }
#>     Pout
#> }
#> <bytecode: 0x000001b4044299a0>
#> <environment: 0x000001b407281af8>
#> 
#> $call
#> glmbfamfunc(family = binomial(logit))
#> 
#> attr(,"class")
#> [1] "glmbfamfunc"

f1<-famfunc$f1
f2<-famfunc$f2
f3<-famfunc$f3
f5<-famfunc$f5
f6<-famfunc$f6