Extract Log-Likelihood

This function is a method function for the "glmb" class used to Extract the log-Likelihood from a Bayesian Generalized Linear Model.

# S3 method for class 'glmb'
logLik(object, ...)

Arguments

object

an object of class glmb, typically the result of a call to glmb

...

further arguments to or from other methods

Value

The function returns a vector, logLikout with the estimated log-likelihood for each draw.

Examples

## ----dobson-------------------------------------------------------------------
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)

## Prior mean vector 
mu<-matrix(0,5)           
mu[1,1]=log(mean(counts)) 
## Prior standard deviation and Variance
mysd<-1           
V=((mysd)^2)*diag(5)  
## Call to glmb
glmb.D93<-glmb(n=1000,counts ~ outcome + treatment,
               family = poisson(),pfamily=dNormal(mu=mu,Sigma=V))
## ----glmb logLik-------------------------------------------------------------
colMeans(logLik(glmb.D93))
#> [1] -25.76712