Prior Family Objects for Bayesian Models

Prior family objects provide a convenient way to specify the details of the priors used by functions such as glmb. See the documentations for lmb, glmb, glmb, and rglmb for the details of how such model fitting takes place.

pfamily(object, ...)

dNormal(mu, Sigma, dispersion = NULL)

dGamma(shape, rate, beta)

dNormal_Gamma(mu, Sigma, shape, rate)

dIndependent_Normal_Gamma(mu, Sigma, shape, rate)

# S3 method for class 'pfamily'
print(x, ...)

Arguments

object

the function pfamily accesses the pfamily objects which are stored within objects created by modelling functions (e.g., glmb).

mu

a prior mean vector for the the modeling coefficients used in several pfamilies

Sigma

a prior Variance-Covariance matrix for the model coefficients in several pfamilies

dispersion

the dispersion to be assumed when it is not given a prior. Should be provided when the Normal prior is for the gaussian(), Gamma(), quasibinomial, or quasipoisson families. The binomial() and poisson() families do not have dispersion coefficients.

shape

the prior shape parameter used by the gamma component of the prior. The gamma distribution is used as a prior for the inverse dispersion coefficients.

rate

the rate parameter used by the gamma component of the prior.

beta

the regression coefficients to be assumed when it is not given a prior. Needs to be provided when the Gamma prior is used for the dispersion. This specification is typically only used as part of Gibbs sampling where the beta and dispersion parameters are updated separately.

x

an object, a pfamily function that is to be printed

...

additional argument(s) for methods.

Value

An object of class "pfamily" (which has a concise print method). This is a list with elements.

pfamily

character: the pfamily name

prior_list

a list with the prior parameters associated with the prior specification

okfamilies

currently implemented families for which the prior family can be used.

plinks

a function that assigns a set of oklinks for the combination of a family and and pfamily.

simfun

function: the function used to generate samples from the posterior density. All currently implemented pfamiles have simulation functions that generate iid samples for the associated posterior distribution.

Details

pfamily is a generic function with methods for classe glmb and lmb. Many glmb models currently only have implementations for the dNormal() prior family. The Gamma() family also works with the dGamma() prior family while the gaussian() family works with the dGamma() and dNormal_Gamma() pfamilies.

References

Cox, D.R. and Snell, E.J. (1981) Applied Statistics; Principles and Examples. London: chapman and Hall.

Dobson, A. J. (1990) An Introduction to Statistical Modeling. London: Chapman and Hall.

Hastie, T. J. and Pregibon, D. (1992) Generalized linear models. Chapter 6 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole. McCullagh P. and Nelder, J. A. (1989) Generalized Linear Models. London: Chapman and Hall.

Nygren, K.N. and Nygren, L.M (2006) Likelihood Subgradient Densities. Journal of the American Statistical Association. vol.101, no.475, pp 1144-1156. doi: 10.1198/016214506000000357.

Raiffa, Howard and Schlaifer, R (1961) Applied Statistical Decision Theory. Boston: Clinton Press, Inc.

See also

Author

The design of the pfamily set of functions was developed by Kjell Nygren and was inspired by the family used by the glmb function to specify the likelihood function. That design in turn was inspired by S functions of the same names described in Hastie and Pregibon (1992).

Examples

mu=c(0,0)
Sigma=diag(2)

npf<-dNormal(mu,Sigma)  # Normal pfamily
str(dNormal(mu,Sigma))
#> List of 6
#>  $ pfamily   : chr "dNormal"
#>  $ prior_list:List of 3
#>   ..$ mu        : num [1:2, 1] 0 0
#>   ..$ Sigma     : num [1:2, 1:2] 1 0 0 1
#>   ..$ dispersion: num 1
#>   ..- attr(*, "Prior Type")= chr "dNormal"
#>  $ okfamilies: chr [1:6] "gaussian" "poisson" "binomial" "quasipoisson" ...
#>  $ plinks    :function (family)  
#>  $ simfun    :function (n, y, x, prior_list, offset = NULL, weights = 1, family = gaussian())  
#>  $ call      : language dNormal(mu = mu, Sigma = Sigma)
#>  - attr(*, "Prior Type")= chr "dNormal"
#>  - attr(*, "class")= chr "pfamily"

## Example where # Normal pfamily is used

## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
mysd<-1
mu<-matrix(0,5)
mu[1]=log(mean(counts))
V0<-((mysd)^2)*diag(5)
glmb.D93<-glmb(counts ~ outcome + treatment, family = poisson(),pfamily=dNormal(mu=mu,Sigma=V0))
summary(glmb.D93)
#> Call
#> glmb(formula = counts ~ outcome + treatment, family = poisson(), 
#>     pfamily = dNormal(mu = mu, Sigma = V0))
#> 
#> Expected Deviance Residuals:
#>        1        2        3        4        5        6        7        8 
#> -0.60482  0.96228 -0.13392 -0.14749 -0.95543  1.09397  0.91757 -0.09613 
#>        9 
#> -0.93447 
#> 
#> Prior and Maximum Likelihood Estimates with Standard Deviations
#> 
#>                          Prior.sd  Max Like. Like.sd
#> (Intercept)  2.813e+00  1.000e+00  3.045e+00   0.171
#> outcome2     0.000e+00  1.000e+00 -4.543e-01   0.202
#> outcome3     0.000e+00  1.000e+00 -2.930e-01   0.193
#> treatment2   0.000e+00  1.000e+00  1.217e-15   0.200
#> treatment3   0.000e+00  1.000e+00  8.438e-16   0.200
#> 
#> Bayesian Estimates Based on 1000 iid draws
#> 
#>              Post.Mode  Post.Mean    Post.Sd MC Error Pr(tail)   
#> (Intercept)  3.027e+00  3.025e+00  1.651e-01        0  0.10889   
#> outcome2    -4.290e-01 -4.402e-01  2.025e-01        0  0.00999 **
#> outcome3    -2.726e-01 -2.882e-01  1.868e-01        0  0.05694 . 
#> treatment2   4.042e-03 -1.138e-03  1.892e-01        0  0.49151   
#> treatment3   4.042e-03 -4.595e-05  1.984e-01        0  0.48452   
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Distribution Percentiles
#> 
#>                  1.0%      2.5%      5.0%    Median     95.0%     97.5%  99.0%
#> (Intercept)  2.654026  2.710085  2.754011  3.027810  3.295622  3.326132  3.377
#> outcome2    -0.943461 -0.859914 -0.779698 -0.430953 -0.106129 -0.052279 -0.008
#> outcome3    -0.738026 -0.670789 -0.619733 -0.284793  0.017850  0.069297  0.107
#> treatment2  -0.421751 -0.360178 -0.309298  0.004976  0.302951  0.355333  0.428
#> treatment3  -0.488597 -0.401604 -0.318289 -0.007816  0.316034  0.392278  0.468
#> 
#> Effective Number of Parameters: 4.827842 
#> Expected Residual Deviance: 9.994012 
#> DIC: 56.45403 
#> 
#> Expected Mean dispersion: 1 
#> Sq.root of Expected Mean dispersion: 1 
#> 
#> Mean Likelihood Subgradient Candidates Per iid sample: 1.793 
#>