Summarizing Bayesian Generalized Linear Model Fits

These functions are all methods for class glmb or summary.glmb objects.

# S3 method for class 'glmb'
summary(object, ...)

# S3 method for class 'summary.glmb'
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object

an object of class "glmb" for which a summary is desired.

x

an object of class "summary.glmb" for which a printed output is desired.

digits

the number of significant digits to use when printing.

...

Additional optional arguments

Value

summary.glmb returns a object of class "summary.glmb", a list with components:

call

the component from object

n

number of draws generated

residuals

vector of mean deviance residuals

coefficients1

Matrix with the prior mean and maximum likelihood coefficients with associated standard deviations

coefficients

Matrix with columns for the posterior mode, posterior mean, posterior standard deviation, monte carlo error, and tail probabilities (posterior probability of observing a value for the coefficient as extreme as the prior mean)

Percentiles

Matrix with estimated percentiles associated with the posterior density

pD

Estimated effective number of parameters

deviance

Vector with draws for the deviance

DIC

Estimated DIC statistic

iters

Average number of candidates per generated draws

Details

The summary.glmb function summarizes the output from the glmb function. Key output includes mean residuals, information related to the prior, mean coefficients with associated stats, percentiles for the coefficients, as well as the effective number of parameters and the DIC statistic.

See also

Examples

###########################  Example for glmb function:
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
print(d.AD <- data.frame(treatment, outcome, counts))
#>   treatment outcome counts
#> 1         1       1     18
#> 2         1       2     17
#> 3         1       3     15
#> 4         2       1     20
#> 5         2       2     10
#> 6         2       3     20
#> 7         3       1     25
#> 8         3       2     13
#> 9         3       3     12

mysd<-1
mu<-matrix(0,5)
mu[1,1]=log(mean(counts))
V0<-((mysd)^2)*diag(5)
glmb.D93<-glmb(counts ~ outcome + treatment, family = poisson(),
pfamily=dNormal(mu=mu,Sigma=V0))
summary(glmb.D93)
#> Call
#> glmb(formula = counts ~ outcome + treatment, family = poisson(), 
#>     pfamily = dNormal(mu = mu, Sigma = V0))
#> 
#> Expected Deviance Residuals:
#>        1        2        3        4        5        6        7        8 
#> -0.55524  1.02000 -0.13518 -0.14375 -0.94014  1.04406  0.94710 -0.05834 
#>        9 
#> -0.95666 
#> 
#> Prior and Maximum Likelihood Estimates with Standard Deviations
#> 
#>                          Prior.sd  Max Like. Like.sd
#> (Intercept)  2.813e+00  1.000e+00  3.045e+00   0.171
#> outcome2     0.000e+00  1.000e+00 -4.543e-01   0.202
#> outcome3     0.000e+00  1.000e+00 -2.930e-01   0.193
#> treatment2   0.000e+00  1.000e+00  1.217e-15   0.200
#> treatment3   0.000e+00  1.000e+00  8.438e-16   0.200
#> 
#> Bayesian Estimates Based on 1000 iid draws
#> 
#>             Post.Mode Post.Mean   Post.Sd MC Error Pr(tail)  
#> (Intercept)  3.027191  3.013257  0.174895        0   0.1229  
#> outcome2    -0.428954 -0.443754  0.197826        0   0.0160 *
#> outcome3    -0.272569 -0.275883  0.188958        0   0.0669 .
#> treatment2   0.004042  0.010060  0.196758        0   0.4995  
#> treatment3   0.004042  0.005572  0.196818        0   0.4985  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Distribution Percentiles
#> 
#>                   1.0%       2.5%       5.0%     Median      95.0%      97.5%
#> (Intercept)  2.6123588  2.6559083  2.7309284  3.0149314  3.2936007  3.3470636
#> outcome2    -0.8680704 -0.8177986 -0.7593678 -0.4477001 -0.1210057 -0.0553648
#> outcome3    -0.6996903 -0.6419489 -0.5916310 -0.2670310  0.0362186  0.1008960
#> treatment2  -0.4685800 -0.3718459 -0.3070046 -0.0001583  0.3327847  0.3818914
#> treatment3  -0.4566942 -0.3577348 -0.3003173  0.0010533  0.3283576  0.4002339
#>             99.0%
#> (Intercept) 3.422
#> outcome2    0.020
#> outcome3    0.174
#> treatment2  0.462
#> treatment3  0.494
#> 
#> Effective Number of Parameters: 4.711922 
#> Expected Residual Deviance: 9.899336 
#> DIC: 56.24344 
#> 
#> Expected Mean dispersion: 1 
#> Sq.root of Expected Mean dispersion: 1 
#> 
#> Mean Likelihood Subgradient Candidates Per iid sample: 1.764 
#> 

###########################  Example for lmb function

## Annette Dobson (1990) "An Introduction to Generalized Linear Models".
## Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)

ps=Prior_Setup(weight ~ group)
#> [1] "Variable names are:"
#> [1] "(Intercept)" "groupTrt"   
mu=ps$mu
V=ps$Sigma
mu[1,1]=mean(weight)

Prior_Check(weight ~ group,family =gaussian(),
            pfamily=dNormal(mu=mu,Sigma=V))
#> [1] "Variables in the Model Are:"
#> [1] "(Intercept)" "groupTrt"   
#> [1] "The maximum likelihood estimates for all coefficients appear to be roughly consistent with the prior."
#>              abs_ratio
#> (Intercept) 0.09242285
#> groupTrt    0.17198871

## May move this step inside the Prior_Check function
lm.D9 <- lm(weight ~ group,x=TRUE,y=TRUE)
disp_ML=sigma(lm.D9)^2
n_prior=2
shape=n_prior/2
rate= disp_ML*shape

lmb.D9=lmb(weight ~ group,dNormal_Gamma(mu,V/disp_ML,shape=shape,rate=rate))
summary(lmb.D9)
#> Call
#> lmb(formula = weight ~ group, pfamily = dNormal_Gamma(mu, V/disp_ML, 
#>     shape = shape, rate = rate))
#> 
#> Expected Deviance Residuals:
#>        1        2        3        4        5        6        7        8 
#> -0.83065  0.57935  0.17935  1.10935 -0.50065 -0.39065  0.16935 -0.47065 
#>        9       10       11       12       13       14       15       16 
#>  0.32935  0.13935  0.13053 -0.50947 -0.26947 -1.08947  1.19053 -0.84947 
#>       17       18       19       20 
#>  1.35053  0.21053 -0.35947  0.01053 
#> 
#> Prior and Maximum Likelihood Estimates with Standard Deviations
#> 
#>             Prior Mean Prior.sd Max Like. Like.sd
#> (Intercept)      4.846    1.436     5.032   0.220
#> groupTrt         0.000    1.436    -0.371   0.311
#> 
#> Bayesian Estimates Based on 1000 iid draws
#> 
#>             Post.Mode Post.Mean Post.Sd MC Error Pr(tail)
#> (Intercept)    5.0081    5.0007  0.2121        0    0.229
#> groupTrt      -0.3311   -0.3212  0.2937        0    0.124
#> 
#> Distribution Percentiles
#> 
#>                1.0%    2.5%    5.0%  Median   95.0%   97.5% 99.0%
#> (Intercept)  4.5175  4.5970  4.6608  4.9892  5.3440  5.4217 5.482
#> groupTrt    -0.9813 -0.8892 -0.8124 -0.3113  0.1722  0.2620 0.352
#> 
#> Effective Number of Parameters: 2.655388 
#> Expected Residual Deviance: 21.45123 
#> DIC: 45.63945 
#> 
#> Expected Mean dispersion: 0.4891574 
#> Sq.root of Expected Mean dispersion: 0.6993979 
#> 
#> Mean Likelihood Subgradient Candidates Per iid sample: 1 
#>